chore: Moves useFileDragging to FileDropzone
